YH-1042: fix: remove keywords & rename entity to resource
