+++ src/entities/resource/model/types.ts
