+++ src/entities/resource/ui/KeywordsListSection/KeywordsListSection.tsx
