+++ src/entities/resource/ui/ResourcesFilterSection/ResourcesFilterSection.module.css
