+++ src/entities/resource/ui/ResourcesFilterSection/ResourcesFilterSection.tsx
